Refactor: EventBookmarkRepository를 BookmarkRepository에 합침
